AI_Agent/requirements.txt
openai>=1.40.0
faiss-cpu
numpy
tiktoken
pyyaml
pydantic>=2
python-dotenv
